Fix build for swift-wasm (#76)
## Description

Hi There, thank you for this library it very cool!

I have a project that targets a server side swift and a web assembly
build and it's currently failing to build for Web Assembly.
Web Assembly uses the same library as linux but comes with a different
OS check so it needs that small fix to compile

Thank you!

## Type of Change

- [X ] Bug fix
- [ ] New feature
- [ ] Breaking change
- [ ] Documentation update

## Additional Notes